requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
seaborn>=0.12.0
h5py==3.15.1
# Optional but highly recommended
pillow>=9.0.0
wandb>=0.15.0
jupyter>=1.0.0
# If you want LPIPS or better FID
torchmetrics>=1.0.0
pytorch-fid>=0.3.0   # for clean FID calculation
